backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
groq==0.4.2
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
jinja2==3.1.3
python-multipart==0.0.6
